Split page layout to component
