+++ framerail/src/lib/types.ts
@@ -8,5 +8,6 @@ export enum Layout {
 }
 export enum PagePane {
   None,
+  Layout,
   Move
 }
