+++ framerail/src/routes/[slug]/[...extra]/index.ts
@@ -1,1 +1,2 @@
+export { default as LayoutPane } from "./LayoutPane.svelte"
 export { default as MovePane } from "./MovePane.svelte"
